refactored getColor function into its own file
